new config file format
The new config file format allows arbitrary resource dependency trees to
be expressed in the config file. Previously, the config file was
specific to Lustre and could not represent other kinds of resources
aside from zpools and Lustre targets.

This also updates the `discover` command to output the config file in
the new format.