test resource agents use CWD when not running interactively
When running interactively, the test resource agents should use the CWD
(generally the root of the source tree) as the location for the state
files, rather than `tests/test_output/{test_id}/` since the latter might
not exist.

However, when running via the test framework, the framework sets the
HALO_TEST_DIRECTORY environment variable to the test-specific private
output directory so that there is no interference with previous runs of
the test agent.