power: use config file to get hosts if not specified
When using the power command and the list of hosts is not specified
on the command linem, assume that the user wants the power status of
every host in the config file.

The config file can either be specified on the command line, or the
command will look in the default location.