client: make discover command support nodeset
When trying to explicitly `discover` many hosts, it can be tedious to
write out all of the hostnames as arguments to the command. These
changes add the capability to pass nodesets to the `discover` command
instead of individual hostnames. Here, a "nodesets" is a string
representation of a set of hostnames in shorthand notation. For example,
the nodeset `node[01,02-04]` represents hosts `node01`, `node02`,
`node03`, and `node04`. This feature is in part inspired by the
[clustershell](https://cea-hpc.github.io/clustershell/) project.