commands: introduce HandledError type for better error handling
Currently, the EmptyError type is being used for error handling by the
command line programs, but it has a problem.

EmptyError can be created from any Error type using the `?` operator,
which makes it to easy to propogate an error as an EmptyError without
handling it.

The programmer using EmptyError must be sure to use inspect_err() or a
similar method before converting an error into an EmptyError, or else
the sysadmin will not get an error message.

It is easy to forget to do this.

The HandledError type resolves this problem because it does not
implement `From<T: std::error::Error>`. The only way to construct a
HandledError, given a `T: Error`, is to use the handle_err() method.

This forces the author of a command line program to ensure that all
error scenarios are handled with an error message, making it harder to
forget about a scenario and get an error without a useful error message.