manager: make manager single threaded
Previously, the manager was "doubly" multithreaded: first, because it
was running in multithreaded tokio runtimes, and second, because 2
runtimes were launched, each started in their own thread. One runtime
was used for the resource management tasks, and one for the CLI server.

This architecture imposes a lot of overhead for no real benefit. For one
thing, the capnproto implementation is necessarily single-threaded [1].
A multithreaded tokio runtime launches extra threads that cannot be used
to send or process RPC requests.

Even if the RPC implementation could take advantage of multiple threads,
it's not clear that it would provide any benefit, since the HALO manager
does not do anything compute intensive.

On the other hand, using a single thread will simplify programming for
the manager.

Previously, because `struct Cluster` was accessed by the resource tasks
and the CLI server in separate threads, all mutable fields of `struct
Cluster` and its children `Host`, `Resource`, etc. required
synchronization.

By running the resource management and CLI server tasks in the same
thread, that synchronization will no longer be necessary.

This will also reduce the footprint of HALO on the host system.

The main negative tradeoff of this change is that it will be important
to ensure that no blocking operations are performed on the single
management thread. For example, any blocking filesystem I/O will
need to be launched on a separate threadpool using
`tokio::task::spawn_blocking()` [2] or a similar technique.

Link: https://github.com/capnproto/capnproto-rust/issues/180#issuecomment-1420847613 [1]
Link: https://docs.rs/tokio/latest/tokio/task/fn.spawn_blocking.html [2]