host: host driven resource management
This introduces a new architecture for managing resources which is
organized primarily at the Host level rather than the ResourceGroup
level.

A task is launched for each Host in the cluster, and this task is
responsible for management of resources that live on that Host. The task
launches child tasks for each ResourceGroup so that multiple
ResourceGroups can be managed concurrently.

Creating each ResourceGroup task as a child of a single Host task makes
it easy to share a single Capnp RPC client (and TCP connection) for all
management actions, allowing more efficient use of resources.

ResourceGroup tasks can be cooperatively canceled by their parent Host
task. This allows a Host task to stop those tasks in a situation where
it must pass responsibility of those resources to its partner Host, for
example, because a failover or failback occurs.

Host tasks communicate with each other using message passing. Host tasks
exchange messages to pass responsibility for managing resources.